On recompile also compile again tsc, also simplified tsc outdir structure
