+++ plugin/compile.js
@@ -2,7 +2,7 @@ const { htmlToJson, jsonToHtml } = require("../compiler")
 const { read, resolve, exists, parse } = require("../util")
 const { PATH_CWD, PATH_DIST } = require("../config")
 const { isEofolTag, compileEofol } = require("./eofol-compile")
-const { clearCompileCache, getCompileCache } = require(resolve(PATH_DIST, "src", "runtime"))
+const { clearCompileCache, getCompileCache } = require(resolve(PATH_DIST, "runtime"))
 
 const importDefs = (scriptPath) => {
   if (exists(scriptPath)) {
@@ -38,7 +38,7 @@ const baseStyles = read(resolve(PATH_CWD, "compiler-data", "styles", "base.css")
 const compile = async (content, filename) => {
   clearCompileCache()
   const parsed = parse(filename)
-  const defs = importDefs(resolve(PATH_DIST, "src", "src", parsed.dir, `${parsed.name}.js`))
+  const defs = importDefs(resolve(PATH_DIST, "src", parsed.dir, `${parsed.name}.js`))
   const json = await htmlToJson(content, false)
   const compiled = compileTree(json, {}, defs)
   const head = compiled.content.find((child) => child.type === "head")
